client synchronisation & store JSON in NVS
When a value is changed by one client, it gets send to all clients and the interface gets adapted, this is working for the dutycycle slider input at the moment. Also saving the driver values for dutycycle in the nvs to later add the functionality to send it to joining clients/saving for reboots. JSON format is updated also